requirements.txt
aiofiles==23.1.0
aiohttp==3.8.4
apache-airflow==2.6.0
beautifulsoup4==4.12.2
numpy==1.21.3
pandas==1.3.4
Requests==2.30.0
schedule==1.2.0
selenium==4.9.0

# To add new dependencies: https://www.youtube.com/watch?v=0UepvC9X4HY
